app/requirements.txt
Flask==3.1.3
psycopg2-binary==2.9.12
requests==2.31.0
